Move Engagement Dashboard API and callback listeners
