--- ee/app/engagement-dashboard/server/listeners/index.js
@@ -1,2 +1,0 @@
-import './messages';
-import './users';
