--- ee/app/engagement-dashboard/server/listeners/messages.js
@@ -1,5 +1,0 @@
-import { callbacks } from '../../../../../app/callbacks/server';
-import { handleMessagesSent, handleMessagesDeleted } from '../lib/messages';
-
-callbacks.add('afterSaveMessage', handleMessagesSent);
-callbacks.add('afterDeleteMessage', handleMessagesDeleted);
