--- ee/app/engagement-dashboard/server/listeners/users.js
@@ -1,4 +1,0 @@
-import { callbacks } from '../../../../../app/callbacks/server';
-import { handleUserCreated } from '../lib/users';
-
-callbacks.add('afterCreateUser', handleUserCreated);
