+++ ee/server/lib/engagementDashboard/users.js
@@ -1,6 +1,6 @@
 import moment from 'moment';
 
-import { Users, Analytics, Sessions } from '../../../../../app/models/server/raw';
+import { Users, Analytics, Sessions } from '../../../../app/models/server/raw';
 import { convertDateToInt, diffBetweenDaysInclusive, getTotalOfWeekItems, convertIntToDate } from './date';
 
 export const handleUserCreated = (user) => {
